commit da documentação + refatoração do código
Refatoração do código.
Adição da documentação do código.
Adição de novos casos de uso.
Correção de bugs.

+++ src/main/java/br/com/coffeework/negocio/service/Service.java
@@ -125,7 +125,10 @@ public abstract class Service<E extends Entidade> implements ServiceFacade<E> {
 
 		final E entidadeobtida = this.obterPorId(entidade.getIdentificador());
 
-		this.getDao().remover(entidadeobtida);
+		if (entidadeobtida != null) {
+
+			this.getDao().remover(entidadeobtida);
+		}
 	}
 
 	/**
